requirements.txt
cryptography==0.8.2
PyJWT==1.1.0
requests==2.6.0
